+++ internal/generate/generator/withgrouped.go
@@ -47,7 +47,7 @@ func generateWithGroupedFunc(code *jen.File, depCount int) {
 			jen.Line().
 				Add(containerParam()),
 			jen.Line().
-				Id("fn").
+				Add(constructorVar()).
 				Func().
 				Params(
 					inputTypes(depCount, imbueContextType())...,
@@ -93,7 +93,7 @@ func generateWithGroupedFuncBody(depCount int, code *jen.Group) {
 							jen.Err(),
 						).
 						Op(":=").
-						Id("fn").
+						Add(constructorVar()).
 						Call(
 							inputVars(depCount, contextVar())...,
 						),