+++ with.gen.go
@@ -5,14 +5,14 @@ package imbue
 // With0 describes how to construct values of type T.
 func With0[T any](
 	con *Container,
-	fn func(*Context) (T, error),
+	ctor func(*Context) (T, error),
 	options ...WithOption,
 ) {
 	t := get[T](con)
 
 	if err := t.Declare(
 		func() (constructor[T], error) {
-			return fn, nil
+			return ctor, nil
 		},
 	); err != nil {
 		panic(err)
@@ -22,7 +22,7 @@ func With0[T any](
 // With1 describes how to construct values of type T from a single dependency.
 func With1[T, D any](
 	con *Container,
-	fn func(*Context, D) (T, error),
+	ctor func(*Context, D) (T, error),
 	options ...WithOption,
 ) {
 	t := get[T](con)
@@ -40,7 +40,7 @@ func With1[T, D any](
 					return v, err
 				}
 
-				return fn(ctx, v1)
+				return ctor(ctx, v1)
 			}, nil
 		},
 	); err != nil {
@@ -51,7 +51,7 @@ func With1[T, D any](
 // With2 describes how to construct values of type T from 2 dependencies.
 func With2[T, D1, D2 any](
 	con *Container,
-	fn func(*Context, D1, D2) (T, error),
+	ctor func(*Context, D1, D2) (T, error),
 	options ...WithOption,
 ) {
 	t := get[T](con)
@@ -79,7 +79,7 @@ func With2[T, D1, D2 any](
 					return v, err
 				}
 
-				return fn(ctx, v1, v2)
+				return ctor(ctx, v1, v2)
 			}, nil
 		},
 	); err != nil {
@@ -90,7 +90,7 @@ func With2[T, D1, D2 any](
 // With3 describes how to construct values of type T from 3 dependencies.
 func With3[T, D1, D2, D3 any](
 	con *Container,
-	fn func(*Context, D1, D2, D3) (T, error),
+	ctor func(*Context, D1, D2, D3) (T, error),
 	options ...WithOption,
 ) {
 	t := get[T](con)
@@ -128,7 +128,7 @@ func With3[T, D1, D2, D3 any](
 					return v, err
 				}
 
-				return fn(ctx, v1, v2, v3)
+				return ctor(ctx, v1, v2, v3)
 			}, nil
 		},
 	); err != nil {
@@ -139,7 +139,7 @@ func With3[T, D1, D2, D3 any](
 // With4 describes how to construct values of type T from 4 dependencies.
 func With4[T, D1, D2, D3, D4 any](
 	con *Container,
-	fn func(*Context, D1, D2, D3, D4) (T, error),
+	ctor func(*Context, D1, D2, D3, D4) (T, error),
 	options ...WithOption,
 ) {
 	t := get[T](con)
@@ -187,7 +187,7 @@ func With4[T, D1, D2, D3, D4 any](
 					return v, err
 				}
 
-				return fn(ctx, v1, v2, v3, v4)
+				return ctor(ctx, v1, v2, v3, v4)
 			}, nil
 		},
 	); err != nil {
@@ -198,7 +198,7 @@ func With4[T, D1, D2, D3, D4 any](
 // With5 describes how to construct values of type T from 5 dependencies.
 func With5[T, D1, D2, D3, D4, D5 any](
 	con *Container,
-	fn func(*Context, D1, D2, D3, D4, D5) (T, error),
+	ctor func(*Context, D1, D2, D3, D4, D5) (T, error),
 	options ...WithOption,
 ) {
 	t := get[T](con)
@@ -256,7 +256,7 @@ func With5[T, D1, D2, D3, D4, D5 any](
 					return v, err
 				}
 
-				return fn(ctx, v1, v2, v3, v4, v5)
+				return ctor(ctx, v1, v2, v3, v4, v5)
 			}, nil
 		},
 	); err != nil {
@@ -267,7 +267,7 @@ func With5[T, D1, D2, D3, D4, D5 any](
 // With6 describes how to construct values of type T from 6 dependencies.
 func With6[T, D1, D2, D3, D4, D5, D6 any](
 	con *Container,
-	fn func(*Context, D1, D2, D3, D4, D5, D6) (T, error),
+	ctor func(*Context, D1, D2, D3, D4, D5, D6) (T, error),
 	options ...WithOption,
 ) {
 	t := get[T](con)
@@ -335,7 +335,7 @@ func With6[T, D1, D2, D3, D4, D5, D6 any](
 					return v, err
 				}
 
-				return fn(ctx, v1, v2, v3, v4, v5, v6)
+				return ctor(ctx, v1, v2, v3, v4, v5, v6)
 			}, nil
 		},
 	); err != nil {
@@ -346,7 +346,7 @@ func With6[T, D1, D2, D3, D4, D5, D6 any](
 // With7 describes how to construct values of type T from 7 dependencies.
 func With7[T, D1, D2, D3, D4, D5, D6, D7 any](
 	con *Container,
-	fn func(*Context, D1, D2, D3, D4, D5, D6, D7) (T, error),
+	ctor func(*Context, D1, D2, D3, D4, D5, D6, D7) (T, error),
 	options ...WithOption,
 ) {
 	t := get[T](con)
@@ -424,7 +424,7 @@ func With7[T, D1, D2, D3, D4, D5, D6, D7 any](
 					return v, err
 				}
 
-				return fn(ctx, v1, v2, v3, v4, v5, v6, v7)
+				return ctor(ctx, v1, v2, v3, v4, v5, v6, v7)
 			}, nil
 		},
 	); err != nil {
@@ -435,7 +435,7 @@ func With7[T, D1, D2, D3, D4, D5, D6, D7 any](
 // With8 describes how to construct values of type T from 8 dependencies.
 func With8[T, D1, D2, D3, D4, D5, D6, D7, D8 any](
 	con *Container,
-	fn func(*Context, D1, D2, D3, D4, D5, D6, D7, D8) (T, error),
+	ctor func(*Context, D1, D2, D3, D4, D5, D6, D7, D8) (T, error),
 	options ...WithOption,
 ) {
 	t := get[T](con)
@@ -523,7 +523,7 @@ func With8[T, D1, D2, D3, D4, D5, D6, D7, D8 any](
 					return v, err
 				}
 
-				return fn(ctx, v1, v2, v3, v4, v5, v6, v7, v8)
+				return ctor(ctx, v1, v2, v3, v4, v5, v6, v7, v8)
 			}, nil
 		},
 	); err != nil {